pyproject.toml
[project]
name = "validation-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "instructor>=1.0",
    "openai>=1.0",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gpu = [
    "pyiqa>=0.1.10",
    "torch>=2.0",
    "transformers>=4.36",
]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
